// src/app/admin/Models/employee.ts
export class Employee {
    emp_id : number;
    emp_name: string;
    emp_username: string;
    emp_email: string;
    emp_password: string;
    emp_phone: string;
    emp_gender: string;
    emp_department: string;
    emp_project: string;
    emp_role: string;
    emp_salary: string;
    emp_doj: string;
    emp_experience: number;
    emp_isAdmin: boolean;

    // constructor(id: number, name: string, department: string, salary: number) {
    //     this.id = id;
    //     this.emp_name = name;
    //     this.emp_department = department;
    //     this.emp_salary = salary;
    // }

    // employee(emp id, name, email, phone, department, role, band, project, salary, DoJ, experience....)

}